+++ routes/productRouter.ts
@@ -1,7 +1,7 @@
 import express from 'express';
 import asyncWrapper from '../middlewares/asyncWrapper';
 import setPagingInfoInReq from '../middlewares/setPagingInfoInReq';
-import { ProductController } from '../src/main/domains/product/controllers/ProductControllers';
+import { ProductController } from '../src/main/domains/product/controllers/ProductController';
 
 const productRouter = express.Router();
 const productController = new ProductController();
